Change volume data save format for optional decimal data received;

+++ 29_Financial/Trial_20_Fmp_API_Parser/List_Comparer/OhlcvObject.cs
@@ -21,7 +21,7 @@ namespace List_Comparer
         public decimal Close { get; set; }
 
         [JsonProperty(PropertyName = "volume")]
-        public int Volume { get; set; }
+        public decimal Volume { get; set; }
         public decimal Capital { get; set; }
         public string Symbol { get; set; }
         public DateTime UtcTimeStamp { get; set; }
